% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.name.R
\name{match.name}
\alias{match.name}
\title{Match original data and geometry data}
\usage{
match.name(data, place, geometry = mapdata, choice = NULL)
}
\arguments{
\item{data}{original data}

\item{place}{place names of original data}

\item{geometry}{geometry data, default is map data with geometry column names of Name}

\item{choice}{a numeric vector, default is NULL}
}
\value{
a dataframe of the original data and geometry data.
}
\description{
Match original data and geometry data
}
