% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\alias{prepare_curve_data}
\alias{prepare_point_data}
\alias{prepare_line_data}
\title{Prepare Data for sf Curve}
\usage{
prepare_curve_data(data)

prepare_point_data(data)

prepare_line_data(data)
}
\arguments{
\item{data}{data with two simple feature columns}
}
\value{
data with xstart, ystart, xend and yend
}
\description{
Prepare Data for sf Curve
}
