% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\alias{px}
\alias{\%px\%}
\alias{ph}
\alias{\%ph\%}
\alias{py}
\alias{\%py\%}
\alias{pv}
\alias{\%pv\%}
\title{Projection}
\usage{
project(sf, k, b, angle = "orth")

px(sf, value = 0)

sf \%px\% value

ph(sf, value)

sf \%ph\% value

py(sf, value)

sf \%py\% value

pv(sf, value)

sf \%pv\% value
}
\arguments{
\item{sf}{as sf object}

\item{k}{projected line slop}

\item{b}{projected line intercept}

\item{angle}{projected direction angle, one numeric or 'orth', default or 'x' or 'y'}

\item{value}{must be a list with length 2}
}
\value{
an MULTIPOINT sf
}
\description{
Projection
}
\examples{
library(sf666)
library(ggplot2)
tri=triangle()
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=project(tri,k=1,b=3))
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=project(tri,k=1,b=3,angle=30))
#project to x-axis
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=px(tri,1))
#project to x-axis
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=tri \%px\% 1)
#project horizontally
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=ph(tri,1))
#project horizontally
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=tri \%ph\% 1)
#project to y-axis
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=py(tri,1))
#project to y-axis
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=tri \%py\% 1)
#project vertically
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=pv(tri,1))
#project vertically
ggplot()+
    geom_sf(data=tri)+
    geom_sf(data=tri \%pv\% 1)
}
