% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gdw.R
\name{read.gdw}
\alias{read.gdw}
\title{read gdw file
gdw is file of getdata graph digitizer}
\usage{
read.gdw(path)
}
\arguments{
\item{path}{path of gdw file}
}
\value{
sf object
}
\description{
read gdw file
gdw is file of getdata graph digitizer
}
