% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.xlsx.R
\name{read.xlsx}
\alias{read.xlsx}
\title{Read xlsx data}
\usage{
read.xlsx(
  path,
  sheet = NULL,
  range = NULL,
  col_names = TRUE,
  col_types = NULL,
  na = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  progress = readxl_progress(),
  .name_repair = "unique"
)
}
\arguments{
\item{path}{see \code{\link[readxl]{read_excel}}}

\item{sheet}{see \code{\link[readxl]{read_excel}}}

\item{range}{see \code{\link[readxl]{read_excel}}}

\item{col_names}{see \code{\link[readxl]{read_excel}}}

\item{col_types}{see \code{\link[readxl]{read_excel}}}

\item{na}{see \code{\link[readxl]{read_excel}}}

\item{trim_ws}{see \code{\link[readxl]{read_excel}}}

\item{skip}{see \code{\link[readxl]{read_excel}}}

\item{n_max}{see \code{\link[readxl]{read_excel}}}

\item{guess_max}{see \code{\link[readxl]{read_excel}}}

\item{progress}{see \code{\link[readxl]{read_excel}}}

\item{.name_repair}{see \code{\link[readxl]{read_excel}}}
}
\value{
a dataframe, for more to see \code{\link[readxl]{read_excel}}
}
\description{
Read xlsx data and return dataframe by package 'readxl'
}
