% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{Recode}
\alias{Recode}
\title{Recode categorical variable}
\usage{
Recode(x, ..., cat = TRUE, to.numeric = FALSE, order = F)
}
\arguments{
\item{x}{dataframe or vector}

\item{...}{code::value}

\item{cat}{logical}

\item{to.numeric}{logcial}

\item{order}{logical}
}
\value{
vector
}
\description{
Recode categorical variable
}
