% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\alias{\%rotate\%}
\title{Rotate}
\usage{
rotate(sf, center = "center2", angel)

sf \%rotate\% center_angel
}
\arguments{
\item{sf}{an sf object}

\item{center}{can be center1 or center2 or any other two-numeric values}

\item{angel}{between -180 and 180}

\item{center_angel}{a list of center and angle, eg. list('center2',30)

ggplot()+
    geom_sf(data = tri)+
    geom_sf(data = tri %rotate% list(c(0,1),90))

#more complicated example
library(ggplot2)
library(sf666)
library(tmcn)
province=mapdata[mapdata$Layer==unique(mapdata$Layer)[1],]
ggplot()+
    geom_sf(data = province,fill='red')+
    geom_sf(data = rotate(province,30,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,60,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,90,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,120,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,150,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,180,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,210,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,240,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,270,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,300,c(80,20)),
            fill='red')+
    geom_sf(data = rotate(province,330,c(80,20)),
            fill='red')+
    theme_void()+
    geom_sf_text(data = txt2,label='Great China',
                 size=13)}
}
\value{
rotated geometry
}
\description{
Rotate two-dimension simple feature.
}
\examples{
library(ggplot2)
library(sf666)
tri=triangle()
ggplot()+
    geom_sf(data = tri)+
    geom_sf(data = rotate(tri,center = c(0,1),angel = 90))
}
