% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_center.R
\name{sf_center}
\alias{sf_center}
\title{Center of sf}
\usage{
sf_center(sf, level = c("1", "2"))
}
\arguments{
\item{sf}{an sf object}

\item{level}{1 or 2. If level equals 1, center of each sf object will be
retruned. If level equals 2, center of all sf object will be returned.}
}
\value{
center coordinate
}
\description{
Center of sf
}
\examples{
library(sf666)
tri=triangle()
sf_center(tri)
}
