% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_line.R
\name{sf_line}
\alias{sf_line}
\title{Create sf Line}
\usage{
sf_line(..., colname = "geometry", rownames, crs = NA_crs_)
}
\arguments{
\item{...}{two-numeric vectors or two-column matrix or list of them}

\item{colname}{colname for simple feature geometry list column}

\item{rownames}{rownames for each sf rows}

\item{crs}{see \code{\link[sf]{sf}}}
}
\value{
an sf object
}
\description{
Create sf Line
}
\details{
sf_line() is a cellection of st_sf(), st_sfc(), st_linestring,
    st_multilinestring() from 'sf' package. The input must be two-numeric
    vectors or two-column matrix or list of them. One independent
    vector or matrix or list will be treated as an sf row. If you want
    to put them together, use list() function for them. The output is
    only sf object. The output has no sfc or sfg.
}
\examples{

library(sf666)

# 1.remeber one vector or one matrix will
# be treated as a row in sf
# 2.remeber one list contain at least two points

# create one line: if only one two-numeric vector is given
# it will be cbinded will c(0,0)
sf_line(c(1,1))

# create two lines: 2 rows
# if two two-numeric vectors are given
# they will be cbinded will c(0,0)
sf_line(c(1,1), c(2,4))

# if you want to combine together
# list() changes the two vectors into a matrix
sf_line(list(c(1,1), c(2,4)))

# create one line: one matrix is a row
mt1=rbind(c(1,1), c(2,4),c(8,9))
sf_line(mt1)

# if you want to put them together, list them
sf_line(list(c(1,2),mt1))

# create many lines by more than one matrix
# one matrix is a row of sf
mt2=rbind(c(1,5),c(2,6),c(5,7))
sf_line(mt1,mt2)

# if you want to put them together, list them
mt3=rbind(c(8,7),c(13,15))
sf_line(list(mt1,mt2),c(4,9),mt3)
}
