% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_point.R
\name{sf_point}
\alias{sf_point}
\title{Create sf Point}
\usage{
sf_point(..., colname = "point", rownames, crs = NA_crs_)
}
\arguments{
\item{...}{two-numeric vectors or two-column matrix or list of them}

\item{colname}{colname for simple feature geometry list column}

\item{rownames}{rownames for each sf rows}

\item{crs}{see \code{\link[sf]{sf}}}
}
\value{
an sf object
}
\description{
Create sf Point
}
\details{
sf_point() is a cellection of st_sf(), st_sfc(), st_point,
    st_multipoint() from 'sf' package. The input must be two-numeric
    vectors or two-column matrix or list of them. One independent
    vector or matrix or list will be treated as an sf row. If you want
    to put them together, use list() function for them. The output is
    only sf object. The output has no sfc or sfg.
}
\examples{
library(sf666)

# remeber one vector or one matrix will
# be treated as a row in sf

# create one point: 1 row
sf_point(c(1,1))

# create two points: 2 rows
sf_point(c(1,1), c(2,4))

# create two points: 1 rows
# list() changes the two vectors into a matrix

sf_point(list(c(1,1), c(2,4)))

# create three points: 2 rows
mt1=rbind(c(1,1), c(2,4))
sf_point(mt1,c(1,2))

# if you want to put them together, list them
sf_point(list(c(1,2),mt1))

# create many points by more than one matrix
# one matrix is a row of sf
mt2=rbind(c(1,5),c(2,6),c(5,7))
sf_point(mt1,mt2)
# if you want to put them together, list them
sf_point(list(mt1,mt2),c(4,9),mt1)
}
