% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_polygon.R
\name{sf_polygon}
\alias{sf_polygon}
\title{Create sf Polygon}
\usage{
sf_polygon(..., colname = "polygon", row.names, crs = NA_agr_)
}
\arguments{
\item{...}{one or more two-column matrix or list of two-column matrix}

\item{colname}{colname for simple feature geometry list column}

\item{row.names}{rownames}

\item{crs}{see \code{\link[sf]{sf}}}
}
\value{
an sf object
}
\description{
Create sf Polygon
}
\details{
sf_polygon() is a cellection of st_sf(), st_sfc(), st_polygon,
    st_multipolygon() from 'sf' package. The input must be matrix or list.
    One independent matrix or list will be treated as a sf row.
    The output is an sf object. The output has no sfc or sfg.
}
\examples{
library(sf666)
# remeber: a matrix, a map, is a row
# if you want to put them together, use list() to treat them into a list

p1 <- rbind(c(0,0), c(1,0), c(3,2), c(2,4), c(1,4), c(0,0))
p2 <- rbind(c(1,1), c(1,2), c(2,2), c(1,1))
# single row
sf_polygon(p1)
# two rows
sf_polygon(p1,p2)
# use list() to combine them together
sf_polygon(list(p1,p2))

p3 <- rbind(c(3,0), c(4,0), c(4,1), c(3,1), c(3,0))
p4 <- rbind(c(3.3,0.3), c(3.8,0.3), c(3.8,0.8), c(3.3,0.8), c(3.3,0.3))[5:1,]
p5 <- rbind(c(3,3), c(4,2), c(4,3), c(3,3))
# a matix, a map, is an af row
# five rows, five maps
sf_polygon(p1,p2,p3,p4,p5)
# put p1 and p2 together: four rows
sf_polygon(list(p1,p2),p3,p4,p5)

# again, put p3 and p4 together: three rows
# the difference from st_multipolygon for sf package is:
# sf_polygon can be given matrix and list, each will be treated as a row
# if you want to put them together, list them again
sf_polygon(list(p1,p2), list(p3,p4), p5)

# again, put list(p1,p2) and list(p3,p4) together in one row
sf_polygon(list(list(p1,p2), list(p3,p4)), p5)
}
