% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_row.R
\name{sf_row}
\alias{sf_row}
\title{Extract Well-Know Text from sf}
\usage{
sf_row(sf, nrow)
}
\arguments{
\item{sf}{an sf object}

\item{nrow}{one or more numbers}
}
\value{
Well-known tex, see \code{\link[sf]{st_as_text}}
}
\description{
Extract Well-Know Text from sf
}
\examples{
sfp=sf_point(c(1,2),c(4,5))
sf_row(sfp,1)
sf_row(sfp,c(1,2))
}
