% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_test.R
\name{sfTest}
\alias{sfTest}
\alias{is.sf}
\alias{is.sfc}
\alias{is.sfg}
\title{Test sf class}
\usage{
is.sf(x)

is.sfc(x)

is.sfg(x)
}
\arguments{
\item{x}{sf, sfc or sfg}
}
\value{
TRUE or FALSE
}
\description{
Test sf class
}
