% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sf_curve.R, R/stat_sf_line.R,
%   R/stat_sf_point.R, R/stat_sf_text.R
\docType{data}
\name{stat_sf_curve}
\alias{stat_sf_curve}
\alias{stat_sf_line}
\alias{StatSfLine}
\alias{prepare_sf_line}
\alias{stat_sf_point}
\alias{StatSfPoints}
\alias{prepare_sf_point}
\alias{stat_sf_text}
\alias{StatSfText}
\title{Create Curve}
\format{
An object of class \code{StatSfLine} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatSfPoints} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatSfText} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
stat_sf_curve(
  mapping = aes(),
  data = NULL,
  CurveData = NULL,
  stat = "identity",
  position = "identity",
  ...,
  curvature = 0.5,
  angle = 90,
  ncp = 5,
  arrow = NULL,
  arrow.fill = NULL,
  lineend = "butt",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_sf_line(
  mapping = aes(),
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

StatSfLine

prepare_sf_line(
  data = NULL,
  mapping = aes(),
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_sf_point(
  mapping = aes(),
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

StatSfPoints

prepare_sf_point(
  data = NULL,
  mapping = aes(),
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_sf_text(
  mapping = aes(),
  data = NULL,
  label = NULL,
  coord_xy = NULL,
  coord_row = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

StatSfText
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{CurveData}{you can give data to this parameter instead of data parameter. Defaulted is NULL.
You can use prepare_curve_data() function to generate this data.}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{curvature}{A numeric value giving the amount of curvature.
    Negative values produce left-hand curves, positive values
    produce right-hand curves, and zero produces a straight line.}

\item{angle}{A numeric value between 0 and 180,
    giving an amount to skew the control
    points of the curve.  Values less than 90 skew the curve towards
    the start point and values greater than 90 skew the curve
    towards the end point.}

\item{ncp}{The number of control points used to draw the curve.
    More control points creates a smoother curve.}

\item{arrow}{specification for arrow heads, as created by \code{\link[grid:arrow]{grid::arrow()}}.}

\item{arrow.fill}{fill colour to use for the arrow head (if closed). \code{NULL}
means use \code{colour} aesthetic.}

\item{lineend}{Line end style (round, butt, square).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
Curve line
}
\description{
Create Curve
}
\examples{
library(sf666)
p1=sf_point(c(1,2))
p2=sf_point(c(2,3))
df=cbind(p1,p2)
library(ggplot2)
ggplot()+
    stat_sf_curve(data=df)


df=cbind(p1,p2,atr=1)
library(ggplot2)
ggplot()+
    stat_sf_curve(data=df,aes(color=atr))
provclass=unique(mapdata$Layer)[1]
prov=mapdata[mapdata$Layer==provclass,]
ggplot(data = prov)+
    geom_sf()+
    stat_sf_point()

# use prepare function to generate a new data for point
# which is a data frame
# so you can adjust the point coord
# then use prepare_sf_point() function to add point
prepare_data=prepare_point_data(prov)
prepare_data$x=prepare_data$x+50
ggplot(data = prov)+
    geom_sf()+
    prepare_sf_point(data=prepare_data)

# example for stat_sf_line
lined=sf_point(c(1,2),c(3,4))
ggplot(lined)+
    stat_sf_line()

ggplot(data = prov)+
    geom_sf()+
    stat_sf_line()


# example for path

prov$bigin=prov$Name[1]
prov$bigincoord=prov$geometry[1]
prov=prov[-1,]
data=prov
prepare_curve_data(prov)
ggplot(prov)+
    geom_sf()+
    stat_sf_curve(data = prov) #data must be give in stat_sf_curve
provclass=unique(mapdata$Layer)[1]
prov=mapdata[mapdata$Layer==provclass,]
ggplot(data = prov)+
    geom_sf()+
    stat_sf_point()

# use prepare function to generate a new data for point
# which is a data frame
# so you can adjust the point coord
# then use prepare_sf_point() function to add point
prepare_data=prepare_point_data(prov)
prepare_data$x=prepare_data$x+50
ggplot(data = prov)+
    geom_sf()+
    prepare_sf_point(data=prepare_data)
}
\keyword{datasets}
