% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{triangle}
\alias{triangle}
\title{Generate Triangle Geometry}
\usage{
triangle(ncol = 1, nrow = 1, w = 1, h = 1)
}
\arguments{
\item{ncol}{number of columns, default is 1}

\item{nrow}{number of rows, default is 1}

\item{w}{weight of each triangle, default is 1}

\item{h}{height of each triangle, default is 1}
}
\value{
triangle geometry
}
\description{
Generate Triangle Geometry
}
\examples{
triangle()
}
