% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\alias{\%z\%}
\alias{zx}
\alias{zoom-towards-x-axis}
\alias{\%zx\%}
\alias{zoom-towards-x-axis-infex}
\alias{zh}
\alias{zoom-horizontal}
\alias{\%zh\%}
\alias{zoom-horizontal-infex}
\alias{zy}
\alias{zoom-towards-y-axis}
\alias{\%zy\%}
\alias{zv}
\alias{zoom-vertical}
\alias{\%zv\%}
\title{Zoom}
\usage{
zoom(sf, ratio = 0.5, center = c(0, 0))

sf \%z\% change

zx(sf, ratio = 1, value = 0)

sf \%zx\% change

zh(sf, ratio = 1, value = 0)

sf \%zh\% change

zy(sf, ratio = 1, value = 0)

sf \%zy\% change

zv(sf, ratio = 1, value = 0)

sf \%zv\% change
}
\arguments{
\item{sf}{an sf object}

\item{ratio}{ratio}

\item{center}{can be center1 or center2 or any other two-numeric values}

\item{change}{a list of ration or center}

\item{value}{value}
}
\value{
a zoomed sf object
}
\description{
Zoom
}
\examples{
library(sf666)
library(ggplot2)
tri=triangle()
ggplot()+
    geom_sf(data = tri)+
    geom_sf(data = zoom(tri,0.5),fill='yellow')
}
